chore(): quick refactor
