--- webx/packages/angular-spa/src/app/enums/field-error.enum.ts
@@ -1,4 +1,0 @@
-export enum FieldError {
-  invalidEmail = 'INVALID_EMAIL',
-  required = 'REQUIRED',
-}
