--- webx/packages/angular-spa/src/app/enums/field-type.enum.ts
@@ -1,4 +1,0 @@
-export enum FieldType {
-  Email = 'EMAIL',
-  Password = 'PASSWORD',
-}
