--- webx/packages/angular-spa/src/app/enums/index.ts
@@ -1,2 +1,0 @@
-export { FieldType } from './field-type.enum';
-export { FieldError } from './field-error.enum';
