--- webx/packages/angular-spa/src/app/interfaces/index.ts
@@ -1,1 +1,0 @@
-export { InputChangedEvent } from './input-changed-event.interface'
