--- webx/packages/react-spa/src/interfaces/index.ts
@@ -1,1 +1,0 @@
-export { InputChangedEvent } from './input-changed-event.interface';
