+++ webx/packages/ui-library/src/components/ui-button/typings/index.ts
@@ -1,0 +1,12 @@
+export type ButtonSize = 'big' | 'small';
+
+export type ButtonTheme = 'primary' | 'secondary';
+
+export type ButtonType = 'button' | 'submit' | 'reset';
+
+export interface UiButtonArgs {
+  disabled: boolean;
+  size: ButtonSize;
+  theme: ButtonTheme;
+  type: ButtonType;
+}
