--- webx/packages/ui-library/src/components/ui-button/ui-button-size.type.ts
@@ -1,1 +1,0 @@
-export type ButtonSize = 'big' | 'small';
