--- webx/packages/ui-library/src/components/ui-button/ui-button-theme.type.ts
@@ -1,1 +1,0 @@
-export type ButtonTheme = 'primary' | 'secondary';
