--- webx/packages/ui-library/src/components/ui-button/ui-button-type.type.ts
@@ -1,1 +1,0 @@
-export type ButtonType = 'button' | 'submit' | 'reset';
