+++ webx/packages/ui-library/src/components/ui-input/constants/index.ts
@@ -1,0 +1,4 @@
+export const ERROR_LABELS = {
+  email: 'Sorry, email seems wrong',
+  required: 'This field is mandatory'
+}
