--- webx/packages/ui-library/src/components/ui-input/interfaces/index.ts
@@ -1,1 +1,0 @@
-export { InputChangedEvent } from './input-changed-event.interface'
