--- webx/packages/ui-library/src/components/ui-input/interfaces/input-changed-event.interface.ts
@@ -1,8 +1,0 @@
-import { FieldType } from "../enums";
-
-export interface InputChangedEvent {
-  target: {
-    value: string;
-    type: FieldType;
-  };
-}
