+++ webx/packages/ui-library/src/components/ui-input/typings/index.ts
@@ -1,0 +1,14 @@
+export type InputType = 'email' | 'password' | 'text' | 'number';
+
+export enum FieldType {
+  Email = 'EMAIL',
+  Password = 'PASSWORD',
+}
+
+export interface InputChangedEvent {
+  target: {
+    value: string;
+    type: FieldType;
+  };
+}
+
