--- webx/packages/ui-library/src/components/ui-input/ui-input-type.type.ts
@@ -1,1 +1,0 @@
-export type InputType = 'email' | 'password' | 'text' | 'number';
