+++ webx/packages/ui-library/src/index.ts
@@ -1,1 +1,3 @@
 export { Components, JSX } from './components';
+export * from './components/ui-button/typings';
+export * from './components/ui-input/typings';
